+++ nefertari_sqla/documents.py
@@ -5,7 +5,8 @@ import six
 from sqlalchemy.orm import (
     class_mapper, object_session, properties, attributes)
 from sqlalchemy.orm.collections import InstrumentedList
-from sqlalchemy.exc import InvalidRequestError, IntegrityError
+from sqlalchemy.exc import (
+    InvalidRequestError, IntegrityError, DataError)
 from sqlalchemy.orm.exc import MultipleResultsFound, NoResultFound
 from sqlalchemy.orm.query import Query
 from sqlalchemy.orm.properties import RelationshipProperty
@@ -16,7 +17,7 @@ from nefertari.json_httpexceptions import (
     JHTTPBadRequest, JHTTPNotFound, JHTTPConflict)
 from nefertari.utils import (
     process_fields, process_limit, _split, dictset,
-    DataProxy)
+    drop_reserved_params)
 from .signals import ESMetaclass, on_bulk_delete
 from .fields import ListField, DictField, IntegerField
 from . import types
@@ -97,22 +98,34 @@ class BaseMixin(object):
             authenticated users.
         _public_fields: String names of fields meant to be displayed to
             non-authenticated users.
+        _hidden_fields: String names of fields meant to be hidden but editable.
         _nested_relationships: String names of relationship fields
             that should be included in JSON data of an object as full
             included documents. If relationship field is not
             present in this list, this field's value in JSON will be an
             object's ID or list of IDs.
+        _nesting_depth: Depth of relationship field nesting in JSON.
+            Defaults to 1(one) which makes only one level of relationship
+            nested.
     """
     _public_fields = None
     _auth_fields = None
+    _hidden_fields = None
     _nested_relationships = ()
+    _nesting_depth = 1
 
     _type = property(lambda self: self.__class__.__name__)
 
     @classmethod
-    def get_es_mapping(cls):
+    def get_es_mapping(cls, _depth=None, types_map=None):
         """ Generate ES mapping from model schema. """
         from nefertari.elasticsearch import ES
+        if types_map is None:
+            types_map = TYPES_MAP
+        if _depth is None:
+            _depth = cls._nesting_depth
+        depth_reached = _depth <= 0
+
         properties = {}
         mapping = {
             ES.src2type(cls.__name__): {
@@ -122,27 +135,28 @@ class BaseMixin(object):
         mapper = class_mapper(cls)
         columns = {c.name: c for c in mapper.columns}
         relationships = {r.key: r for r in mapper.relationships}
-        # Replace field 'id' with primary key field
-        columns['id'] = columns.get(cls.pk_field())
 
         for name, column in columns.items():
             column_type = column.type
             if isinstance(column_type, types.ChoiceArray):
                 column_type = column_type.impl.item_type
             column_type = type(column_type)
-            if column_type not in TYPES_MAP:
+            if column_type not in types_map:
                 continue
-            properties[name] = TYPES_MAP[column_type]
+            properties[name] = types_map[column_type]
 
         for name, column in relationships.items():
-            if name in cls._nested_relationships:
-                column_type = {'type': 'object'}
+            if name in cls._nested_relationships and not depth_reached:
+                column_type = {'type': 'nested'}
+                submapping = column.mapper.class_.get_es_mapping(
+                    _depth=_depth-1)
+                column_type.update(list(submapping.values())[0])
             else:
                 rel_pk_field = column.mapper.class_.pk_field_type()
-                column_type = TYPES_MAP[rel_pk_field]
+                column_type = types_map[rel_pk_field]
             properties[name] = column_type
 
-        properties['_type'] = {'type': 'string'}
+        properties['_pk'] = {'type': 'string'}
         return mapping
 
     @classmethod
@@ -208,6 +222,8 @@ class BaseMixin(object):
                 fields_only = [
                     f for f in fields_only if f not in fields_exclude]
             if fields_only:
+                # Add PK field
+                fields_only.append(cls.pk_field())
                 fields_only = [
                     getattr(cls, f) for f in sorted(set(fields_only))]
                 query_set = query_set.with_entities(*fields_only)
@@ -237,28 +253,25 @@ class BaseMixin(object):
     def filter_objects(cls, objects, first=False, **params):
         """ Perform query with :params: on instances sequence :objects:
 
-        Arguments:
-            :object: Sequence of :cls: instances on which query should be run.
-            :params: Query parameters.
+        :param object: Sequence of :cls: instances on which query should be run.
+        :param params: Query parameters to filter :objects:.
         """
         id_name = cls.pk_field()
         ids = [getattr(obj, id_name, None) for obj in objects]
         ids = [str(id_) for id_ in ids if id_ is not None]
         field_obj = getattr(cls, id_name)
 
-        session = Session()
-        query_set = session.query(cls).filter(field_obj.in_(ids))
+        query_set = Session().query(cls).filter(field_obj.in_(ids))
 
-        if first:
-            params['_limit'] = 1
-            params['__raise_on_empty'] = True
+        if params:
             params['query_set'] = query_set.from_self()
             query_set = cls.get_collection(**params)
 
+        if first:
             first_obj = query_set.first()
             if not first_obj:
-                msg = "'{}({}={})' resource not found".format(
-                    cls.__name__, id_name, params[id_name])
+                msg = "'{}({})' resource not found".format(
+                    cls.__name__, params)
                 raise JHTTPNotFound(msg)
             return first_obj
 
@@ -304,14 +317,82 @@ class BaseMixin(object):
 
     @classmethod
     def get_collection(cls, **params):
-        """
-        Params may include '_limit', '_page', '_sort', '_fields'.
-        Returns paginated and sorted query set.
-        Raises JHTTPBadRequest for bad values in params.
+        """ Query collection and return results.
+
+        Notes:
+        *   Before validating that only model fields are present in params,
+            reserved params, query params and all params starting with
+            double underscore are dropped.
+        *   Params which have value "_all" are dropped.
+        *   When ``_count`` param is used, objects count is returned
+            before applying offset and limit.
+
+        :param bool __strict: If True ``params`` are validated to contain
+            only fields defined on model, exception is raised if invalid
+            fields are present. When False - invalid fields are dropped.
+            Defaults to ``True``.
+        :param bool _item_request: Indicates whether it is a single item
+            request or not. When True and DataError happens on DB request,
+            JHTTPNotFound is raised. JHTTPBadRequest is raised when False.
+            Defaults to ``False``.
+        :param list _sort: Field names to sort results by. If field name
+            is prefixed with "-" it is used for "descending" sorting.
+            Otherwise "ascending" sorting is performed by that field.
+            Defaults to an empty list in which case sorting is not
+            performed.
+        :param list _fields: Names of fields which should be included
+            or excluded from results. Fields to excluded should be
+            prefixed with "-". Defaults to an empty list in which
+            case all fields are returned.
+        :param int _limit: Number of results per page. Defaults
+            to None in which case all results are returned.
+        :param int _page: Number of page. In conjunction with
+            ``_limit`` is used to calculate results offset. Defaults to
+            None in which case it is ignored. Params ``_page`` and
+            ``_start` are mutually exclusive.
+        :param int _start: Results offset. If provided ``_limit`` and
+            ``_page`` params are ignored when calculating offset. Defaults
+            to None. Params ``_page`` and ``_start`` are mutually
+            exclusive. If not offset-related params are provided, offset
+            equals to 0.
+        :param Query query_set: Existing queryset. If provided, all queries
+            are applied to it instead of creating new queryset. Defaults
+            to None.
+        :param _count: When provided, only results number is returned as
+            integer.
+        :param _explain: When provided, query performed(SQL) is returned
+            as a string instead of query results.
+        :param bool __raise_on_empty: When True JHTTPNotFound is raised
+            if query returned no results. Defaults to False in which case
+            error is just logged and empty query results are returned.
+
+        :returns: Query results as ``sqlalchemy.orm.query.Query`` instance.
+            May be sorted, offset, limited.
+        :returns: Dict of {'field_name': fieldval}, when ``_fields`` param
+            is provided.
+        :returns: Number of query results as an int when ``_count`` param
+            is provided.
+        :returns: String representing query ran when ``_explain`` param
+            is provided.
+
+        :raises JHTTPNotFound: When ``__raise_on_empty=True`` and no
+            results found.
+        :raises JHTTPNotFound: When ``_item_request=True`` and
+            ``sqlalchemy.exc.DataError`` exception is raised during DB
+            query. Latter exception is raised when querying DB with
+            an identifier of a wrong type. E.g. when querying Int field
+            with a string.
+        :raises JHTTPBadRequest: When ``_item_request=False`` and
+            ``sqlalchemy.exc.DataError`` exception is raised during DB
+            query.
+        :raises JHTTPBadRequest: When ``sqlalchemy.exc.InvalidRequestError``
+            or ``sqlalchemy.exc.IntegrityError`` errors happen during DB
+            query.
         """
         log.debug('Get collection: {}, {}'.format(cls.__name__, params))
         params.pop('__confirmation', False)
         __strict = params.pop('__strict', True)
+        _item_request = params.pop('_item_request', False)
 
         _sort = _split(params.pop('_sort', []))
         _fields = _split(params.pop('_fields', []))
@@ -337,6 +418,7 @@ class BaseMixin(object):
 
         iterables_exprs, params = cls._pop_iterables(params)
 
+        params = drop_reserved_params(params)
         if __strict:
             _check_fields = [
                 f.strip('-+') for f in list(params.keys()) + _fields + _sort]
@@ -362,16 +444,14 @@ class BaseMixin(object):
             if _count:
                 return _total
 
-            if _limit is None:
-                raise JHTTPBadRequest('Missing _limit')
-
-            _start, _limit = process_limit(_start, _page, _limit)
-
             # Filtering by fields has to be the first thing to do on
             # the query_set!
             query_set = cls.apply_fields(query_set, _fields)
             query_set = cls.apply_sort(query_set, _sort)
-            query_set = query_set.offset(_start).limit(_limit)
+
+            if _limit is not None:
+                _start, _limit = process_limit(_start, _page, _limit)
+                query_set = query_set.offset(_start).limit(_limit)
 
             if not query_set.count():
                 msg = "'%s(%s)' resource not found" % (cls.__name__, params)
@@ -380,8 +460,15 @@ class BaseMixin(object):
                 else:
                     log.debug(msg)
 
-        except (InvalidRequestError,) as e:
-            raise JHTTPBadRequest(str(e), extra={'data': e})
+        except DataError as ex:
+            if _item_request:
+                msg = "'{}({})' resource not found".format(
+                    cls.__name__, params)
+                raise JHTTPNotFound(msg, explanation=ex.message)
+            else:
+                raise JHTTPBadRequest(str(ex), extra={'data': ex})
+        except (InvalidRequestError,) as ex:
+            raise JHTTPBadRequest(str(ex), extra={'data': ex})
 
         query_sql = str(query_set).replace('\n', '')
         if _explain:
@@ -389,12 +476,37 @@ class BaseMixin(object):
 
         log.debug('get_collection.query_set: %s (%s)', cls.__name__, query_sql)
 
+        if _fields:
+            query_set = cls.add_field_names(query_set, _fields)
+
         query_set._nefertari_meta = dict(
             total=_total,
             start=_start,
             fields=_fields)
-
         return query_set
+
+    @classmethod
+    def add_field_names(cls, query_set, requested_fields):
+        """ Convert list of tuples to dict with proper field keys. """
+        from .utils import FieldsQuerySet
+        fields = [col['name'] for col in query_set.column_descriptions] + [
+            '_type']
+        add_vals = (cls.__name__,)
+        pk_field = cls.pk_field()
+
+        def _convert(val):
+            return dict(zip(fields, val+add_vals))
+
+        def _add_pk(obj):
+            if pk_field in obj:
+                obj['_pk'] = obj[pk_field]
+                if pk_field not in requested_fields:
+                    obj.pop(pk_field)
+            return obj
+
+        values = query_set.all()
+        converted = [_add_pk(_convert(val)) for val in values]
+        return FieldsQuerySet(converted)
 
     @classmethod
     def has_field(cls, field):
@@ -402,10 +514,17 @@ class BaseMixin(object):
 
     @classmethod
     def native_fields(cls):
-        mapper = class_mapper(cls)
-        columns = [c.name for c in mapper.columns]
-        relationships = [r.key for r in mapper.relationships]
+        columns = list(cls._mapped_columns().keys())
+        relationships = list(cls._mapped_relationships().keys())
         return columns + relationships
+
+    @classmethod
+    def _mapped_columns(cls):
+        return {c.name: c for c in class_mapper(cls).columns}
+
+    @classmethod
+    def _mapped_relationships(cls):
+        return {c.key: c for c in class_mapper(cls).relationships}
 
     @classmethod
     def fields_to_query(cls):
@@ -414,16 +533,19 @@ class BaseMixin(object):
         return list(set(query_fields + cls.native_fields()))
 
     @classmethod
-    def get_resource(cls, **params):
+    def get_item(cls, **params):
+        """ Get single item and raise exception if not found.
+
+        Exception raising when item is not found can be disabled
+        by passing ``__raise_on_empty=False`` in params.
+
+        :returns: Single collection item as an instance of ``cls``.
+        """
         params.setdefault('__raise_on_empty', True)
         params['_limit'] = 1
+        params['_item_request'] = True
         query_set = cls.get_collection(**params)
         return query_set.first()
-
-    @classmethod
-    def get(cls, **kw):
-        return cls.get_resource(
-            __raise_on_empty=kw.pop('__raise', False), **kw)
 
     def unique_fields(self):
         native_fields = class_mapper(self.__class__).columns
@@ -465,13 +587,13 @@ class BaseMixin(object):
         return self
 
     @classmethod
-    def _delete_many(cls, items, request_params=None,
+    def _delete_many(cls, items, request=None,
                      synchronize_session=False):
         """ Delete :items: queryset or objects list.
 
-        When queryset passed, Query.delete() is used to delete it. Note that
-        queryset may not have limit(), offset(), order_by(), group_by(), or
-        distinct() called on it.
+        When queryset passed, Query.delete() is used to delete it but
+        first queryset is re-queried to clean it from explicit
+        limit/offset/etc.
 
         If some of the methods listed above were called, or :items: is not
         a Query instance, one-by-one items update is performed.
@@ -482,51 +604,60 @@ class BaseMixin(object):
         'after_bulk_delete' ORM event.
         """
         if isinstance(items, Query):
-            items_count = cls.count(items)
-            try:
-                delete_items = items.all()
-                items.delete(
-                    synchronize_session=synchronize_session)
-                on_bulk_delete(cls, delete_items, request_params)
-                return items_count
-            except Exception as ex:
-                log.error(str(ex))
+            del_queryset = cls._clean_queryset(items)
+            del_items = del_queryset.all()
+            del_count = del_queryset.delete(
+                synchronize_session=synchronize_session)
+            on_bulk_delete(cls, del_items, request)
+            return del_count
         items_count = len(items)
         session = Session()
         for item in items:
-            item._request_params = request_params
+            item._request = request
             session.delete(item)
         session.flush()
         return items_count
 
     @classmethod
-    def _update_many(cls, items, params, request_params=None,
+    def _update_many(cls, items, params, request=None,
                      synchronize_session='fetch'):
         """ Update :items: queryset or objects list.
 
-        When queryset passed, Query.update() is used to update it. Note that
-        queryset may not have limit(), offset(), order_by(), group_by(), or
-        distinct() called on it.
+        When queryset passed, Query.update() is used to update it but
+        first queryset is re-queried to clean it from explicit
+        limit/offset/etc.
 
         If some of the methods listed above were called, or :items: is not
         a Query instance, one-by-one items update is performed.
         """
         if isinstance(items, Query):
-            items._request_params = request_params
-            items.update(
+            upd_queryset = cls._clean_queryset(items)
+            upd_queryset._request = request
+            upd_count = upd_queryset.update(
                 params, synchronize_session=synchronize_session)
-            return cls.count(items)
+            return upd_count
         items_count = len(items)
         for item in items:
-            item.update(params, request_params)
+            item.update(params, request)
         return items_count
 
+    @classmethod
+    def _clean_queryset(cls, queryset):
+        """ Clean :queryset: from explicit limit, offset, etc.
+
+        New queryset is created by querying collection by IDs from
+        passed queryset.
+        """
+        pk_field = getattr(cls, cls.pk_field())
+        pks_query = queryset.with_entities(pk_field)
+        return queryset.session.query(cls).filter(
+            pk_field.in_(pks_query))
+
     def __repr__(self):
-        parts = []
-
-        if hasattr(self, 'id'):
-            parts.append('id=%s' % self.id)
-
+        pk_field = self.pk_field()
+        parts = [
+            '{}={}'.format(pk_field, getattr(self, pk_field)),
+        ]
         if hasattr(self, '_version'):
             parts.append('v=%s' % self._version)
 
@@ -541,11 +672,13 @@ class BaseMixin(object):
     @classmethod
     def get_null_values(cls):
         """ Get null values of :cls: fields. """
+        skip_fields = {'_version', '_acl'}
         null_values = {}
-        mapper = class_mapper(cls)
-        columns = {c.name: c for c in mapper.columns}
-        columns.update({r.key: r for r in mapper.relationships})
+        columns = cls._mapped_columns()
+        columns.update(cls._mapped_relationships())
         for name, col in columns.items():
+            if name in skip_fields:
+                continue
             if isinstance(col, RelationshipProperty) and col.uselist:
                 value = []
             else:
@@ -554,27 +687,38 @@ class BaseMixin(object):
         return null_values
 
     def to_dict(self, **kwargs):
+        _depth = kwargs.get('_depth')
+        if _depth is None:
+            _depth = self._nesting_depth
+        depth_reached = _depth is not None and _depth <= 0
+
+        _data = dictset()
         native_fields = self.__class__.native_fields()
-        _data = {}
         for field in native_fields:
             value = getattr(self, field, None)
+
             include = field in self._nested_relationships
-            if not include:
-                get_id = lambda v: getattr(v, v.pk_field(), None)
-                if isinstance(value, BaseMixin):
-                    value = get_id(value)
-                elif isinstance(value, InstrumentedList):
-                    value = [get_id(val) for val in value]
+            if not include or depth_reached:
+                encoder = lambda v: getattr(v, v.pk_field(), None)
+            else:
+                encoder = lambda v: v.to_dict(_depth=_depth-1)
+
+            if isinstance(value, BaseMixin):
+                value = encoder(value)
+            elif isinstance(value, InstrumentedList):
+                value = [encoder(val) for val in value]
+            elif hasattr(value, 'to_dict'):
+                value = value.to_dict(_depth=_depth-1)
+
             _data[field] = value
-        _dict = DataProxy(_data).to_dict(**kwargs)
-        _dict['_type'] = self._type
-        _dict['id'] = getattr(self, self.pk_field())
-        return _dict
+        _data['_type'] = self._type
+        _data['_pk'] = str(getattr(self, self.pk_field()))
+        return _data
 
     def update_iterables(self, params, attr, unique=False,
                          value_type=None, save=True,
-                         request_params=None):
-        self._request_params = request_params
+                         request=None):
+        self._request = request
         mapper = class_mapper(self.__class__)
         columns = {c.name: c for c in mapper.columns}
         is_dict = isinstance(columns.get(attr), DictField)
@@ -612,7 +756,7 @@ class BaseMixin(object):
 
             setattr(self, attr, final_value)
             if save:
-                self.save(request_params)
+                self.save(request)
 
         def update_list(update_params):
             final_value = getattr(self, attr, []) or []
@@ -641,26 +785,43 @@ class BaseMixin(object):
 
             setattr(self, attr, final_value)
             if save:
-                self.save(request_params)
+                self.save(request)
 
         if is_dict:
             update_dict(params)
         elif is_list:
             update_list(params)
 
-    def get_reference_documents(self):
-        # TODO: Make lazy load of documents
+    def get_related_documents(self, nested_only=False):
+        """ Return pairs of (Model, istances) of relationship fields.
+
+        Pair contains of two elements:
+          :Model: Model class object(s) contained in field.
+          :instances: Model class instance(s) contained in field
+
+        :param nested_only: Boolean, defaults to False. When True, return
+            results only contain data for models on which current model
+            and field are nested.
+        """
         iter_props = class_mapper(self.__class__).iterate_properties
         backref_props = [p for p in iter_props
                          if isinstance(p, properties.RelationshipProperty)]
+
         for prop in backref_props:
             value = getattr(self, prop.key)
-            # Do not index empty values and 'Many' side in OneToMany,
-            # when 'One' side is indexed.
-            # If 'Many' side should be indexed, its value is already a list.
-            if value is None or isinstance(value, list):
+            # Do not index empty values
+            if not value:
                 continue
-            yield (value.__class__, [value.to_dict()])
+            if not isinstance(value, list):
+                value = [value]
+            model_cls = value[0].__class__
+
+            if nested_only:
+                backref = prop.back_populates
+                if backref and backref not in model_cls._nested_relationships:
+                    continue
+
+            yield (model_cls, value)
 
     def _is_modified(self):
         """ Determine if instance is modified.
@@ -677,6 +838,10 @@ class BaseMixin(object):
                 if history.added or history.deleted:
                     return True
 
+    def _is_created(self):
+        state = attributes.instance_state(self)
+        return not state.persistent
+
 
 class BaseDocument(BaseObject, BaseMixin):
     """ Base class for SQLA models.
@@ -692,17 +857,15 @@ class BaseDocument(BaseObject, BaseMixin):
         if self._is_modified():
             self._version = (self._version or 0) + 1
 
-    def save(self, request_params=None):
+    def save(self, request=None):
         session = object_session(self)
         self._bump_version()
-        self._request_params = request_params
+        self._request = request
         session = session or Session()
         try:
-            self.apply_before_validation()
             session.add(self)
             session.flush()
             session.expire(self)
-            self.apply_after_validation()
             return self
         except (IntegrityError,) as e:
             if 'duplicate' not in e.args[0]:
@@ -713,16 +876,14 @@ class BaseDocument(BaseObject, BaseMixin):
                     self.__class__.__name__),
                 extra={'data': e})
 
-    def update(self, params, request_params=None):
-        self._request_params = request_params
+    def update(self, params, request=None):
+        self._request = request
         try:
             self._update(params)
             self._bump_version()
-            self.apply_before_validation()
             session = object_session(self)
             session.add(self)
             session.flush()
-            self.apply_after_validation()
             return self
         except (IntegrityError,) as e:
             if 'duplicate' not in e.args[0]:
@@ -733,63 +894,17 @@ class BaseDocument(BaseObject, BaseMixin):
                     self.__class__.__name__),
                 extra={'data': e})
 
-    def delete(self, request_params=None):
-        self._request_params = request_params
+    def delete(self, request=None):
+        self._request = request
         object_session(self).delete(self)
 
-    def apply_processors(self, column_names=None, before=False, after=False):
-        """ Apply processors to columns with :column_names: names.
-
-        Arguments:
-          :column_names: List of string names of changed columns.
-          :before: Boolean indicating whether to apply before_validation
-            processors.
-          :after: Boolean indicating whether to apply after_validation
-            processors.
-        """
-        columns = {c.key: c for c in class_mapper(self.__class__).columns}
-        if column_names is None:
-            column_names = columns.keys()
-
-        for name in column_names:
-            column = columns.get(name)
-            if column is not None and hasattr(column, 'apply_processors'):
-                new_value = getattr(self, name)
-                processed_value = column.apply_processors(
-                    instance=self, new_value=new_value,
-                    before=before, after=after)
-                if new_value != processed_value:
-                    setattr(self, name, processed_value)
-
-    def apply_before_validation(self):
-        """ Determine changed columns and run `self.apply_processors` to
-        apply needed processors.
-
-        Note that at this stage, field values are in the exact same state
-        you posted/set them. E.g. if you set time_field='11/22/2000',
-        self.time_field will be equal to '11/22/2000' here.
-        """
-        columns = {c.key: c for c in class_mapper(self.__class__).columns}
-        state = attributes.instance_state(self)
-
-        if state.persistent:
-            changed_columns = list(state.committed_state.keys())
-        else:  # New object
-            changed_columns = list(columns.keys())
-
-        changed_columns = sorted(changed_columns)
-        self._columns_to_process = changed_columns
-        self.apply_processors(changed_columns, before=True)
-
-    def apply_after_validation(self):
-        """ Run `self.apply_processors` with columns names determined by
-        `self.apply_before_validation`.
-
-        Note that at this stage, field values are in the exact same state
-        you posted/set them. E.g. if you set time_field='11/22/2000',
-        self.time_field will be equal to '11/22/2000' here.
-        """
-        self.apply_processors(self._columns_to_process, after=True)
+    @classmethod
+    def get_field_params(cls, field_name):
+        """ Get init params of column named :field_name:. """
+        columns = cls._mapped_columns()
+        columns.update(cls._mapped_relationships())
+        column = columns.get(field_name)
+        return getattr(column, '_init_kwargs', None)
 
 
 class ESBaseDocument(six.with_metaclass(ESMetaclass, BaseDocument)):