+++ nefertari_sqla/tests/test_documents.py
@@ -74,6 +74,9 @@ class TestBaseMixin(object):
     def test_get_es_mapping(self, memory_db):
         class MyModel(docs.BaseDocument):
             __tablename__ = 'mymodel'
+            _nested_relationships = ['parent']
+            _nesting_depth = 0
+
             my_id = fields.IdField()
             name = fields.StringField(primary_key=True)
             settings = fields.DictField()
@@ -84,6 +87,8 @@ class TestBaseMixin(object):
         class MyModel2(docs.BaseDocument):
             _nested_relationships = ['myself']
             __tablename__ = 'mymodel2'
+            _nesting_depth = 1
+
             name = fields.StringField(primary_key=True)
             myself = fields.Relationship(
                 document='MyModel', backref_name='parent',
@@ -93,29 +98,35 @@ class TestBaseMixin(object):
                 ref_column_type=fields.StringField)
         memory_db()
 
-        assert MyModel.get_es_mapping() == {
-            'mymodel': {
+        my_model_mapping = MyModel.get_es_mapping()
+
+        assert my_model_mapping == {
+            'MyModel': {
                 'properties': {
-                    '_type': {'type': 'string'},
+                    '_pk': {'type': 'string'},
                     '_version': {'type': 'long'},
                     'settings': {'type': 'object', 'enabled': False},
                     'groups': {'type': 'string'},
-                    'id': {'type': 'string'},
                     'my_id': {'type': 'long'},
                     'name': {'type': 'string'},
                     'parent': {'type': 'string'}
                 }
             }
         }
-        assert MyModel2.get_es_mapping() == {
-            'mymodel2': {
+        my_model2_mapping = MyModel2.get_es_mapping()
+        myself_props = my_model_mapping['MyModel']['properties']
+
+        assert my_model2_mapping == {
+            'MyModel2': {
                 'properties': {
-                    '_type': {'type': 'string'},
+                    '_pk': {'type': 'string'},
                     '_version': {'type': 'long'},
                     'child_id': {'type': 'string'},
-                    'id': {'type': 'string'},
                     'name': {'type': 'string'},
-                    'myself': {'type': 'object'}
+                    'myself': {
+                        'type': 'nested',
+                        'properties': myself_props
+                    }
                 }
             }
         }
@@ -203,7 +214,7 @@ class TestBaseMixin(object):
         _fields = ['title']
         MyModel.apply_fields(query_set, _fields)
         query_set.with_entities.assert_called_once_with(
-            MyModel.title)
+            MyModel.id, MyModel.title)
 
     def test_apply_fields_no_any_fields(self, simple_model, memory_db):
         memory_db()
@@ -238,23 +249,34 @@ class TestBaseMixin(object):
         assert count == 12345
 
     @patch.object(docs, 'Session')
-    @patch.object(docs.BaseMixin, 'get_collection')
-    def test_filter_objects(
-            self, mock_get, mock_sess, simple_model, memory_db):
+    def test_filter_objects_first(
+            self, mock_sess, simple_model, memory_db):
         memory_db()
-        queryset1 = mock_sess().query().filter()
-        queryset2 = Mock()
-        mock_get.return_value = queryset2
         simple_model.id.in_ = Mock()
-        simple_model.filter_objects([Mock(id=4)], first=True)
+        result = simple_model.filter_objects([Mock(id=4)], first=True)
 
         mock_sess().query.assert_called_with(simple_model)
-        assert mock_sess().query().filter.call_count == 2
-
-        mock_get.assert_called_once_with(
-            _limit=1, __raise_on_empty=True,
-            query_set=queryset1.from_self())
+        assert mock_sess().query().filter.call_count == 1
         simple_model.id.in_.assert_called_once_with(['4'])
+        assert result == mock_sess().query().filter().first()
+
+    @patch.object(docs, 'Session')
+    @patch.object(docs.BaseMixin, 'get_collection')
+    def test_filter_objects_params(
+            self, mock_get, mock_sess, simple_model, memory_db):
+        memory_db()
+        simple_model.id.in_ = Mock()
+        result = simple_model.filter_objects(
+            [Mock(id=4), Mock(id=5)], foo='bar')
+
+        mock_sess().query.assert_called_with(simple_model)
+        assert mock_sess().query().filter.call_count == 1
+        simple_model.id.in_.assert_called_once_with(['4', '5'])
+        query_set = mock_sess().query().filter()
+        mock_get.assert_called_once_with(
+            query_set=query_set.from_self(),
+            foo='bar')
+        assert result == mock_get()
 
     def test_pop_iterables(self, memory_db):
         class MyModel(docs.BaseDocument):
@@ -277,6 +299,33 @@ class TestBaseMixin(object):
         expected = 'DictField database querying is not supported'
         assert str(ex.value) == expected
 
+    def test_add_field_names_no_pk_requested(
+            self, memory_db, simple_model):
+        from sqlalchemy.orm.query import Query
+        memory_db()
+        simple_model(id=1, name='foo').save()
+        queryset = simple_model.get_collection(_limit=1)
+        queryset = queryset.with_entities(
+            simple_model.name, simple_model.id)
+        assert isinstance(queryset, Query)
+        objects = simple_model.add_field_names(queryset, [])
+        assert objects == [{'_type': 'MyModel', 'name': 'foo', '_pk': 1}]
+
+    def test_add_field_names_pk_requested(
+            self, memory_db, simple_model):
+        from sqlalchemy.orm.query import Query
+        memory_db()
+        simple_model(id=1, name='foo').save()
+        queryset = simple_model.get_collection(_limit=1)
+        queryset = queryset.with_entities(
+            simple_model.name, simple_model.id)
+        assert isinstance(queryset, Query)
+        objects = simple_model.add_field_names(queryset, ['id'])
+        assert objects == [{
+            '_type': 'MyModel', 'name': 'foo', '_pk': 1,
+            'id': 1
+        }]
+
     @patch.object(docs.BaseMixin, 'native_fields')
     def test_has_field(self, mock_fields):
         mock_fields.return_value = ['foo', 'bar']
@@ -284,31 +333,40 @@ class TestBaseMixin(object):
         assert not docs.BaseMixin.has_field('bazz')
 
     @patch.object(docs.BaseMixin, 'get_collection')
-    def test_get_resource(self, mock_get_coll):
+    def test_get_item(self, mock_get_coll):
         queryset = Mock()
         mock_get_coll.return_value = queryset
-        resource = docs.BaseMixin.get_resource(foo='bar')
+        resource = docs.BaseMixin.get_item(foo='bar')
         mock_get_coll.assert_called_once_with(
-            __raise_on_empty=True, _limit=1, foo='bar')
+            __raise_on_empty=True, _limit=1, foo='bar',
+            _item_request=True)
         mock_get_coll().first.assert_called_once_with()
         assert resource == mock_get_coll().first()
 
     def test_native_fields(self, simple_model, memory_db):
         memory_db()
-        assert simple_model.native_fields() == [
+        assert sorted(simple_model.native_fields()) == [
             '_version', 'id', 'name']
+
+    def test_mapped_columns(self, simple_model, memory_db):
+        memory_db()
+        cols = simple_model._mapped_columns().keys()
+        assert sorted(cols) == ['_version', 'id', 'name']
+
+    @patch.object(docs, 'class_mapper')
+    def test_mapped_relationships(
+            self, mock_mapper, simple_model, memory_db):
+        memory_db()
+        rels = [Mock(key='foo')]
+        mock_mapper.return_value = Mock(relationships=rels)
+        cols = simple_model._mapped_relationships()
+        assert cols == {'foo': rels[0]}
 
     def test_fields_to_query(self, simple_model, memory_db):
         memory_db()
         assert sorted(simple_model.fields_to_query()) == [
             '_count', '_fields', '_limit', '_page', '_sort',
             '_start', '_version', 'id', 'name']
-
-    @patch.object(docs.BaseMixin, 'get_resource')
-    def test_get(self, get_res):
-        docs.BaseMixin.get(foo='bar')
-        get_res.assert_called_once_with(
-            __raise_on_empty=False, foo='bar')
 
     def test_unique_fields(self, memory_db):
         class MyModel(docs.BaseDocument):
@@ -384,37 +442,47 @@ class TestBaseMixin(object):
         mock_session().flush.assert_called_once_with()
 
     @patch.object(docs, 'on_bulk_delete')
-    def test_underscore_delete_many_query(self, mock_on_bulk):
+    @patch.object(docs.BaseMixin, '_clean_queryset')
+    def test_underscore_delete_many_query(
+            self, mock_clean, mock_on_bulk):
         from sqlalchemy.orm.query import Query
         items = Query('asd')
-        items.count = Mock(return_value=4)
-        items.all = Mock(return_value=[1, 2, 3])
-        items.delete = Mock()
-        assert docs.BaseMixin._delete_many(items) == 4
-        items.delete.assert_called_once_with(
+        clean_items = Query("ASD")
+        clean_items.all = Mock(return_value=[1, 2, 3])
+        clean_items.delete = Mock()
+        mock_clean.return_value = clean_items
+        count = docs.BaseMixin._delete_many(items)
+        mock_clean.assert_called_once_with(items)
+        clean_items.delete.assert_called_once_with(
             synchronize_session=False)
         mock_on_bulk.assert_called_once_with(
             docs.BaseMixin, [1, 2, 3], None)
+        assert count == clean_items.delete()
 
     def test_underscore_update_many(self):
         item = Mock()
         assert docs.BaseMixin._update_many([item], {'foo': 'bar'}) == 1
         item.update.assert_called_once_with({'foo': 'bar'}, None)
 
-    def test_underscore_update_many_query(self):
+    @patch.object(docs.BaseMixin, '_clean_queryset')
+    def test_underscore_update_many_query(self, mock_clean):
         from sqlalchemy.orm.query import Query
         items = Query('asd')
-        items.update = Mock()
-        items.count = Mock(return_value=4)
-        assert docs.BaseMixin._update_many(items, {'foo': 'bar'}) == 4
-        items.update.assert_called_once_with(
+        clean_items = Query("ASD")
+        clean_items.all = Mock(return_value=[1, 2, 3])
+        clean_items.update = Mock()
+        mock_clean.return_value = clean_items
+        count = docs.BaseMixin._update_many(items, {'foo': 'bar'})
+        mock_clean.assert_called_once_with(items)
+        clean_items.update.assert_called_once_with(
             {'foo': 'bar'}, synchronize_session='fetch')
+        assert count == clean_items.update()
 
-    def test_repr(self):
-        obj = docs.BaseMixin()
+    def test_repr(self, simple_model, memory_db):
+        obj = simple_model()
         obj.id = 3
         obj._version = 12
-        assert str(obj) == '<BaseMixin: id=3, v=12>'
+        assert str(obj) == '<MyModel: id=3, v=12>'
 
     @patch.object(docs.BaseMixin, 'get_collection')
     def test_get_by_ids(self, mock_coll, memory_db):
@@ -444,14 +512,12 @@ class TestBaseMixin(object):
                 document='MyModel1', backref_name='model2')
 
         assert MyModel1.get_null_values() == {
-            '_version': None,
             'fk_field': None,
             'name': None,
             'model2': None,
         }
 
         assert MyModel2.get_null_values() == {
-            '_version': None,
             'models1': [],
             'name': None,
         }
@@ -472,9 +538,11 @@ class TestBaseMixin(object):
 
         result = myobj1.to_dict()
         assert list(sorted(result.keys())) == [
-            '_type', '_version', 'id', 'other_obj', 'other_obj2', 'other_obj3']
+            '_pk', '_type', '_version', 'id', 'other_obj',
+            'other_obj2', 'other_obj3']
         assert result['_type'] == 'MyModel'
         assert result['id'] == 1
+        assert result['_pk'] == '1'
         # Not nester one-to-one
         assert result['other_obj'] == 2
         # Not nester many-to-one
@@ -483,6 +551,19 @@ class TestBaseMixin(object):
         assert isinstance(result['other_obj3'], dict)
         assert result['other_obj3']['_type'] == 'MyModel'
         assert result['other_obj3']['id'] == 4
+
+    def test_to_dict_depth(self, memory_db):
+        class MyModel(docs.BaseDocument):
+            __tablename__ = 'mymodel'
+            _nested_relationships = ['other_obj']
+            id = fields.IdField(primary_key=True)
+            other_obj = fields.StringField()
+        memory_db()
+        myobj1 = MyModel(id=1)
+        myobj1.other_obj = MyModel(id=2)
+
+        result = myobj1.to_dict(_depth=0)
+        assert result['other_obj'] == 2
 
     @patch.object(docs, 'object_session')
     def test_update_iterables_dict(self, obj_session, memory_db):
@@ -556,7 +637,7 @@ class TestBaseMixin(object):
         myobj.update_iterables("", attr='settings', unique=False)
         assert myobj.settings == []
 
-    def test_get_reference_documents(self, memory_db):
+    def test_get_related_documents(self, memory_db):
 
         class Child(docs.BaseDocument):
             __tablename__ = 'child'
@@ -573,16 +654,31 @@ class TestBaseMixin(object):
 
         memory_db()
 
+        # Item
         parent = Parent(id=1)
         child = Child(id=1, parent=parent)
-        result = [v for v in child.get_reference_documents()]
+        result = [v for v in child.get_related_documents()]
         assert len(result) == 1
         assert result[0][0] is Parent
-        assert result[0][1] == [parent.to_dict()]
+        assert result[0][1] == [parent]
 
-        # 'Many' side of relationship values are not returned
+        # Collection
         assert child in parent.children
-        result = [v for v in parent.get_reference_documents()]
+        result = [v for v in parent.get_related_documents()]
+        assert len(result) == 1
+        assert result[0][0] is Child
+        assert result[0][1] == [child]
+
+        # nested_only=True for nested object
+        Child._nested_relationships = ('parent',)
+        result = [v for v in parent.get_related_documents(nested_only=True)]
+        assert len(result) == 1
+        assert result[0][0] is Child
+        assert result[0][1] == [child]
+
+        # nested_only=True for NOT nested object
+        Parent._nested_relationships = ()
+        result = [v for v in child.get_related_documents(nested_only=True)]
         assert len(result) == 0
 
     def test_is_modified_id_not_persistent(self, memory_db, simple_model):
@@ -598,16 +694,23 @@ class TestBaseMixin(object):
     def test_is_modified_same_value_set(self, memory_db, simple_model):
         memory_db()
         obj = simple_model(id=1, name='foo').save()
-        obj = simple_model.get(id=1)
+        obj = simple_model.get_item(id=1)
         obj.name = 'foo'
         assert not obj._is_modified()
 
     def test_is_modified(self, memory_db, simple_model):
         memory_db()
         obj = simple_model(id=1, name='foo').save()
-        obj = simple_model.get(id=1)
+        obj = simple_model.get_item(id=1)
         obj.name = 'bar'
         assert obj._is_modified()
+
+    def test_is_created(self, memory_db, simple_model):
+        memory_db()
+        obj = simple_model(id=1, name='foo')
+        assert obj._is_created()
+        obj = obj.save()
+        assert not obj._is_created()
 
 
 class TestBaseDocument(object):
@@ -680,92 +783,6 @@ class TestBaseDocument(object):
         obj_session.assert_called_once_with(obj)
         obj_session().delete.assert_called_once_with(obj)
 
-    def test_apply_before_validation_new_object(self, memory_db):
-        processor = lambda instance, new_value: 'foobar'
-        processor2 = lambda instance, new_value: new_value + '+'
-
-        class MyModel(docs.BaseDocument):
-            __tablename__ = 'mymodel'
-            id = fields.IdField(primary_key=True)
-            name = fields.StringField(
-                before_validation=[processor],
-                after_validation=[processor2])
-            email = fields.StringField(before_validation=[processor])
-        memory_db()
-
-        obj = MyModel(name='myname')
-        obj.apply_before_validation()
-        assert obj.name == 'foobar'
-        assert obj.email == 'foobar'
-
-    def test_apply_before_validation_existing_object(self, memory_db):
-        processor = lambda instance, new_value: new_value + '-'
-        processor2 = lambda instance, new_value: new_value + '+'
-
-        class MyModel(docs.BaseDocument):
-            __tablename__ = 'mymodel'
-            id = fields.IdField(primary_key=True)
-            name = fields.StringField(
-                before_validation=[processor],
-                after_validation=[processor2])
-            email = fields.StringField(before_validation=[processor])
-        memory_db()
-
-        obj = MyModel(id=1, name='myname', email='FOO').save()
-        assert obj.name == 'myname-+'
-        assert obj.email == 'FOO-'
-
-        obj = MyModel.get(id=1)
-        assert obj.name == 'myname-+'
-        obj.name = 'supername'
-        obj.apply_before_validation()
-        assert obj.name == 'supername-'
-        assert obj.email == 'FOO-'
-
-    def test_apply_after_validation(self, memory_db):
-        memory_db()
-        obj = docs.BaseDocument()
-        obj.apply_processors = Mock()
-        obj._columns_to_process = [1, 2, 3]
-        obj.apply_after_validation()
-        obj.apply_processors.assert_called_once_with(
-            [1, 2, 3], after=True)
-
-    def test_apply_before_validation(self, memory_db):
-        class MyModel(docs.BaseDocument):
-            __tablename__ = 'mymodel'
-            id = fields.IdField(primary_key=True)
-        memory_db()
-
-        obj = MyModel(id=1)
-        obj.apply_processors = Mock()
-        obj.apply_before_validation()
-        obj.apply_processors.assert_called_once_with(
-            ['_version', 'id'], before=True)
-
-    def test_apply_processors(self, memory_db):
-        class MyModel(docs.BaseDocument):
-            __tablename__ = 'mymodel'
-            name = fields.StringField(
-                primary_key=True,
-                before_validation=[lambda instance, new_value: new_value + '-'],
-                after_validation=[
-                    lambda instance, new_value: new_value + '+'])
-        memory_db()
-        obj = MyModel(name='foo')
-
-        obj.apply_processors(before=True)
-        assert obj.name == 'foo-'
-
-        obj.apply_processors(after=True)
-        assert obj.name == 'foo-+'
-
-        obj.apply_processors()
-        assert obj.name == 'foo-+'
-
-        obj.apply_processors(column_names=['name'], before=True, after=True)
-        assert obj.name == 'foo-+-+'
-
 
 class TestGetCollection(object):
 
@@ -814,7 +831,7 @@ class TestGetCollection(object):
         memory_db()
         simple_model(id=1, name='foo').save()
         result = simple_model.get_collection(_limit=1, _fields=['name'])
-        assert result.all() == [('foo',)]
+        assert result == [{'_pk': 1, '_type': 'MyModel', 'name': 'foo'}]
 
     def test_offset(self, simple_model, memory_db):
         memory_db()
@@ -862,6 +879,20 @@ class TestGetCollection(object):
         except JHTTPNotFound:
             raise Exception('Unexpected JHTTPNotFound exception')
 
+    @patch.object(docs, 'drop_reserved_params')
+    @patch.object(docs.BaseMixin, 'check_fields_allowed')
+    def test_reserved_params_dropped(
+            self, mock_check, mock_drop, simple_model, memory_db):
+        from nefertari.utils import dictset
+        memory_db()
+        mock_drop.side_effect = lambda x: dictset({'name': 'a'})
+        try:
+            simple_model.get_collection(_limit=1, __strict=True)
+        except JHTTPBadRequest:
+            raise Exception('Unexpected JHTTPBadRequest exception')
+        mock_drop.assert_called_once_with({})
+        mock_check.assert_called_once_with(['name'])
+
     def test_queryset_metadata(self, simple_model, memory_db):
         memory_db()
         simple_model(id=1, name='foo').save()
@@ -869,8 +900,3 @@ class TestGetCollection(object):
         assert queryset._nefertari_meta['total'] == 1
         assert queryset._nefertari_meta['start'] == 0
         assert queryset._nefertari_meta['fields'] == []
-
-    def test_no_limit(self, simple_model, memory_db):
-        memory_db()
-        with pytest.raises(JHTTPBadRequest):
-            simple_model.get_collection()